Create Dino Game
Creating a simple Dino game in C programming can be a fun project. Below is a basic console-based Dino game source code. This code allows the player (the dinosaur) to jump over obstacles by pressing the spacebar key.

Please note that this is a simplified version of the Dino game, and you can enhance it with additional features and graphics.